Start refurbishing the scalar properties
