+++ utils/mrisurf_vals.h
@@ -53,9 +53,12 @@ static int compare_sort_vals(const void *pc1, const void *pc2)
 
 
 
-int mrisClearMomentum(MRI_SURFACE *mris);
 
+int mrisClearMomentum(MRI_SURFACE *mris);
+int mrisClearGradient(MRI_SURFACE *mris);
 int mrisClearExtraGradient(MRI_SURFACE *mris);
+
+void mrisSetVal(MRIS *mris, float val);
 
 
 void MRISsetCurvaturesToValues    (MRIS *mris, int fno);    //  val = curv
@@ -63,6 +66,7 @@ void MRISsetCurvaturesToOrigValues(MRIS *mris, int fno);    //  etc.
 void MRISsetOrigValuesToCurvatures(MRIS *mris, int fno);
 void MRISsetOrigValuesToValues    (MRIS *mris, int fno);
 void MRISsetValuesToCurvatures    (MRIS *mris, int fno);    
+
 int detectContrast(MRIS *mris);
 
 int mrisUpdateTargetLocations(MRI_SURFACE *mris, MRI *mri, double target_intensity);
